Move demo mode module to core
