+++ app/auth/auth-android/build.gradle.kts
@@ -8,6 +8,6 @@ dependencies {
   implementation(libs.koin.android)
   implementation(projects.authCore)
   implementation(projects.coreCommonPublic)
-  implementation(projects.featureDemoMode)
+  implementation(projects.coreDemoMode)
   implementation(projects.navigationActivity)
 }
