+++ app/data/data-forever/src/main/kotlin/com/hedvig/android/data/forever/di/ForeverRepositoryProvider.kt
@@ -2,7 +2,7 @@ package com.hedvig.android.data.forever.di
 
 import com.hedvig.android.data.forever.ForeverRepository
 import com.hedvig.android.feature.demomode.DemoManager
-import com.hedvig.android.feature.demomode.ProdOrDemoProvider
+import com.hedvig.android.core.demomode.ProdOrDemoProvider
 
 class ForeverRepositoryProvider(
   demoManager: DemoManager,
