--- app/feature/feature-demo-mode/build.gradle.kts
@@ -1,5 +1,0 @@
-plugins {
-  id("hedvig.android.ktlint")
-  id("hedvig.android.library")
-  alias(libs.plugins.squareSortDependencies)
-}
