--- app/feature/feature-demo-mode/src/main/kotlin/com/hedvig/android/feature/demomode/DemoManager.kt
@@ -1,5 +1,0 @@
-package com.hedvig.android.feature.demomode
-
-class DemoManager() {
-  val isDemoMode = true
-}
