--- app/feature/feature-demo-mode/src/main/kotlin/com/hedvig/android/feature/demomode/Provider.kt
@@ -1,5 +1,0 @@
-package com.hedvig.android.feature.demomode
-
-fun interface Provider<T> {
-  fun provide(): T
-}
