+++ app/feature/feature-forever/build.gradle.kts
@@ -40,8 +40,8 @@ dependencies {
   implementation(projects.coreUi)
   implementation(projects.coreUiData)
   implementation(projects.dataForever)
-  implementation(projects.featureDemoMode)
   implementation(projects.languageCore)
+  implementation(projects.coreDemoMode)
   implementation(projects.moleculeAndroid)
   implementation(projects.moleculePublic)
   implementation(projects.navigationComposeTyped)
