+++ app/feature/feature-home/build.gradle.kts
@@ -51,8 +51,8 @@ dependencies {
   implementation(projects.dataTravelCertificatePublic)
   implementation(projects.hanalyticsCore)
   implementation(projects.hanalyticsFeatureFlagsPublic)
-  implementation(projects.featureDemoMode)
   implementation(projects.languageCore)
+  implementation(projects.coreDemoMode)
   implementation(projects.marketCore)
   implementation(projects.memberRemindersPublic)
   implementation(projects.memberRemindersUi)
