+++ app/feature/feature-home/src/main/kotlin/com/hedvig/android/feature/home/di/GetHomeDataUseCaseProvider.kt
@@ -1,7 +1,7 @@
 package com.hedvig.android.feature.home.di
 
 import com.hedvig.android.feature.demomode.DemoManager
-import com.hedvig.android.feature.demomode.ProdOrDemoProvider
+import com.hedvig.android.core.demomode.ProdOrDemoProvider
 import com.hedvig.android.feature.home.home.data.GetHomeDataUseCase
 
 internal class GetHomeDataUseCaseProvider(
