+++ app/feature/feature-insurances/build.gradle.kts
@@ -43,8 +43,7 @@ dependencies {
   implementation(projects.navigationCore)
   implementation(projects.notificationBadgeDataPublic)
   implementation(projects.pullrefresh)
-  implementation(projects.featureDemoMode)
-
+  implementation(projects.coreDemoMode)
   testImplementation(libs.assertK)
   testImplementation(libs.coroutines.test)
   testImplementation(libs.junit)
