+++ app/feature/feature-insurances/src/main/kotlin/com/hedvig/android/feature/insurances/di/GetCrossSellsUseCaseProvider.kt
@@ -1,7 +1,7 @@
 package com.hedvig.android.feature.insurances.di
 
 import com.hedvig.android.feature.demomode.DemoManager
-import com.hedvig.android.feature.demomode.ProdOrDemoProvider
+import com.hedvig.android.core.demomode.ProdOrDemoProvider
 import com.hedvig.android.feature.insurances.data.GetCrossSellsUseCase
 
 internal class GetCrossSellsUseCaseProvider(
