+++ app/feature/feature-profile/src/main/kotlin/com/hedvig/android/feature/profile/di/ProfileRepositoryProvider.kt
@@ -1,7 +1,7 @@
 package com.hedvig.android.feature.profile.di
 
 import com.hedvig.android.feature.demomode.DemoManager
-import com.hedvig.android.feature.demomode.ProdOrDemoProvider
+import com.hedvig.android.core.demomode.ProdOrDemoProvider
 import com.hedvig.android.feature.profile.data.ProfileRepository
 
 internal class ProfileRepositoryProvider(
