+++ app/notification/notification-firebase/build.gradle.kts
@@ -18,6 +18,6 @@ dependencies {
   implementation(projects.authCore)
   implementation(projects.authEventCore)
   implementation(projects.coreCommonPublic)
-  implementation(projects.featureDemoMode)
+  implementation(projects.coreDemoMode)
   implementation(projects.notificationCore)
 }
